refactor: move codelyzer rules to the correct folder
