+++ src/converters/lintConfigs/rules/ruleConverters/codelyzer/import-destructuring-spacing.ts
@@ -1,4 +1,4 @@
-import { RuleConverter } from "../ruleConverter";
+import { RuleConverter } from "../../ruleConverter";
 
 export const convertImportDestructuringSpacing: RuleConverter = () => {
     return {};
