+++ src/converters/lintConfigs/rules/ruleConverters/codelyzer/tests/prefer-inline-decorator.test.ts
